+++ modules/task_2/brazhnik_d_sort_shell_simple_merge_omp/CMakeLists.txt
@@ -1,7 +1,7 @@
 get_filename_component(ProjectId ${CMAKE_CURRENT_SOURCE_DIR} NAME)
 
-if ( USE_SEQ )
-    set(ProjectId "${ProjectId}_seq")
+if ( USE_OMP )
+    set(ProjectId "${ProjectId}_omp")
     project( ${ProjectId} )
     message( STATUS "-- " ${ProjectId} )
 
@@ -52,6 +52,6 @@ if ( USE_SEQ )
             ${ProjectId}
             ${ARGS_FOR_CHECK_COUNT_TESTS}
     )
-else( USE_SEQ )
+else( USE_OMP )
     message( STATUS "-- ${ProjectId} - NOT BUILD!"  )
-endif( USE_SEQ )
+endif( USE_OMP )
