+++ modules/task_2/brazhnik_d_sort_shell_simple_merge_omp/short_shell.h
@@ -1,7 +1,7 @@
 // Copyright 2022 Brazhnik Dmitry
 
-#ifndef MODULES_TASK_2_BRAZHNIK_D_SORT_SHELL_SIMPLE_MERGE_SHORT_SHELL_H_
-#define MODULES_TASK_2_BRAZHNIK_D_SORT_SHELL_SIMPLE_MERGE_SHORT_SHELL_H_
+#ifndef MODULES_TASK_2_BRAZHNIK_D_SORT_SHELL_SIMPLE_MERGE_OMP_SHORT_SHELL_H_
+#define MODULES_TASK_2_BRAZHNIK_D_SORT_SHELL_SIMPLE_MERGE_OMP_SHORT_SHELL_H_
 
 #include <vector>
 #include <utility>
@@ -14,4 +14,4 @@ std::vector<int> Merge_of_two_vectors(const std::vector<int>& a, const std::vect
 std::vector<int> Shell_with_merge(const std::vector<int>& a, int n, int size);
 std::vector<int> Shell_with_merge_omp(const std::vector<int>& a, int n, int size);
 
-#endif  // MODULES_TASK_2_BRAZHNIK_D_SORT_SHELL_SIMPLE_MERGE_SHORT_SHELL_H_
+#endif  // MODULES_TASK_2_BRAZHNIK_D_SORT_SHELL_SIMPLE_MERGE_OMP_SHORT_SHELL_H_
